Enemy spawning
Fixed some merge issues that arose from the scene not being properly saved so the enemy now behaves as intended.
Added in an enemy spawner and set the enemy as a prefab. Alongside this made a game manager which handles whenever a new wave of enemies need to be spawned in.